Added a more DE-agnostic version.

+++ upscreenshot.sh
@@ -1,24 +1,20 @@
 #!/bin/bash
 
-NAME=$(mktemp --dry-run XXXXX.jpg)
+NAME=$(mktemp --dry-run XXXXX.png)
 SITE="i.brod.es"
 USER="aaron"
 DIR="/var/www/images/$NAME"
-PORT=22
-DE="kde4"
-
-convert $@ /tmp/$NAME
-
-scp -P $PORT /tmp/$NAME $USER@$SITE:$DIR
-
-rm /tmp/$NAME
-
 URL="http://$SITE/$NAME"
+PORT=22
 
-if [$DE eq "kde4"]; then
-  qdbus org.kde.klipper /klipper setClipboardContents "$URL"
-  kdialog --passivepopup "Uploaded" 1
-elif [$DE eq "xfce4"]; then
-
-
+if [ ! -z $1 ]; then
+  convert $@ /tmp/$NAME
+  out=`scp -P $PORT /tmp/$NAME $USER@$SITE:$DIR`
+  rm /tmp/$NAME
+  xsel < "$URL"
+  if [ $out == 0 ]; then
+    notify-send 'Uploaded and pasted into clipboard.' $URL --icon=dialog-information
+  else
+    notify-send 'Upload failed.' --icon=dialog-error
+  fi
 fi
